// src/app/menu/model/models.ts
export interface Pizza {
    name: string;
    description: string;
    photoUrl: string;
    rating: string;
}

export interface Sandwich {
    name: string;
    description: string;
    photoUrl: string;
    rating: string;
}

export interface Drink {
    name: string;
    description: string;
    photoUrl: string;
    rating: string;
}

export interface Dessert {
    name: string;
    description: string;
    photoUrl: string;
    rating: string;
}